Update garment.js
changed the way artwork gets added to the master file. instead of just placing one piece of artwork per graphic, script now loops each graphic location and attempts to find artwork guides on the master file. if found, script will scale the art to match the size of the guide and make a copy of the artwork to align with the guide.