bug fixes
runtime error when there is no adult garment because the youth art was being copied from the adult art. adjusted this so that there's a generic "logo art" variable and then the adult and youth art variables are made from that.

Turned off app messages during saving to prevent "can't embed this font" messages that need to be manually clicked away.

Added a library for easily getting the file name of an FDS garment because some of them are formatted differently from how the builder outputs them.