logo text fix
fixed logo text logic. if there was a space in the middle of the existing logo, the character case was incorrectly rendered for the rest of the logo. added logic to accomodate a space and still use the right case for the rest of the logo. also formatted garment object.