bug fix
bug fix for when there's a discrepancy between layer name and garment code. in this example, the data showed "BM-10001", but the appropriate layer was actually called "FD-10001". Stripped out the initial letters and hyphen or underscore when looking for parent layer to prevent this kind of issue. Also added logic to check for a case when no garment layer is found and alert the error rather than fail spectacularly with a non-helpful error message.