optimize graphic styles
added logic to audit the color data to "flatten" any color styles that would appear as a solid color. For example, if a C1 color should be replaced with a Royal Blue background and a Royal Blue top to bottom gradient. A royal blue gradient on top of a royal blue background will just look royal blue. So to save time, this would be flattened to a simple fill color.