disable pattern flattening
logic that ignores pattern fills if they are the same color as the background has been disabled. There are certain pattern fills that are visible even when they are the same fill color as the background, need to figure out which ones those are so that we can conditionally disable this based on the pattern.